fix(redux/account):  fixed delete button
