+++ client/components/index.js
@@ -10,3 +10,4 @@ export {default as Draw} from './draw'
 export {default as UserAccount} from './account'
 export {default as FormAddImage} from './FormAddImage'
 export {default as SingleImage} from './SingleImage'
+export {default as ImageHistory} from './ImageHistory'
